// archive/solid-chat/types/chat.ts
type observerConfigType = (
  accessor: string | boolean | object | MessageType[],
) => void
export type observersConfigType = Record<
  'observeUserInput' | 'observeLoading' | 'observeMessages',
  observerConfigType
>

export type BotProps = {
  botId: string
  chatId?: string
  apiHost?: string
  onRequest?: (request: RequestInit) => Promise<void>
  chatflowConfig?: Record<string, unknown>
  backgroundColor?: string
  welcomeMessage?: string
  errorMessage?: string
  botMessage?: BotMessageTheme
  userMessage?: UserMessageTheme
  textInput?: TextInputTheme
  feedback?: FeedbackTheme
  bubbleBackgroundColor?: string
  bubbleTextColor?: string
  showTitle?: boolean
  showAgentMessages?: boolean
  title?: string
  titleAvatarSrc?: string
  titleTextColor?: string
  titleBackgroundColor?: string
  formBackgroundColor?: string
  formTextColor?: string
  fontSize?: number
  isFullPage?: boolean
  footer?: FooterTheme
  sourceDocsTitle?: string
  observersConfig?: observersConfigType
  starterPrompts?: string[] | Record<string, { prompt: string }>
  starterPromptFontSize?: number
  clearChatOnReload?: boolean
  disclaimer?: DisclaimerPopUpTheme
  dateTimeToggle?: DateTimeToggleTheme
  renderHTML?: boolean
  closeBot?: () => void
}